fix(ci): isolate latency-sensitive #1167 attach-width tests from contended full-suite + add dedicated job
The three #1167 attach-width regression tests
(TestStartAttachPTY_FullWidthFromFrameOne, TestStartAttachPTY_MatchesNarrowTerminal
in internal/tmux; TestRemoteAttach_FullWidthFromFrameOne in internal/session)
depend on the tmux server completing client window-size arbitration after an
attach. In release.yml's `go test -race ./...` on a 4-vCPU ubuntu-latest runner,
the attach client/server subprocesses are CPU-starved past the test's deadline
by the dozens of co-scheduled race-instrumented package binaries, so
`#{window_width}` is still read at tmux's 80-col birth default and the tests
fail at the full deadline (~5.0s). Root cause is cross-package CPU starvation,
not a tmux version skew, socket collision, or sub-deadline race.

These tests also had ZERO PR coverage: no workflow runs ./internal/tmux on PRs,
and session-persistence.yml filters internal/session to -run TestPersistence_.
The only place they executed was the release full-suite, so a "green" PR meant
they never ran.

Fix (assertion unchanged — never forces the size via resize-window):
- Gate the three width-arbitration tests behind a `tmux_timing` build tag so
  they are excluded from the default `./...` build. The two tmux tests move to a
  new issue1167_attach_width_timing_test.go; the fast failure-mode test
  (FallsBackWhenSizeUnavailable) stays in the default build, keeping PR coverage
  of the size-probe fallback path.
- Harden the wait: first poll `tmux list-clients #{client_width}` until the
  client registers at the controlling-terminal width (proves the client side is
  done), then poll `#{window_width}`, with a 30s load-proportional deadline
  (mirrors the cgroup-isolation test budget). This stops mistaking "subprocess
  not scheduled yet" for a regression without weakening the assertion.
- Add .github/workflows/issue1167-attach-width.yml: a dedicated, low-contention
  job that runs only these three tests, isolated, with -p 1, on PRs touching
  internal/tmux or internal/session and on release tags — giving #1167 real PR
  coverage with no cross-package starvation.

Verified: `go test -race ./internal/tmux/ ./internal/session/` (default tags)
passes with the timing tests excluded; `go test -tags tmux_timing -p 1` runs and
passes all three in isolation (~0.04s each); golangci-lint clean for both tag
sets.

Refs #1167

Committed by Ashesh Goplani
